persistencia de oficina com repository do spring

+++ src/main/java/br/com/eadtt/aula01/controller/request/DeleteOficinaFilterRequest.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 @Data
 public class DeleteOficinaFilterRequest {
-    private List<Long> ids;
+    private List<Integer> ids;
     private String nome;
 }
